Creating the resource not found exception. The repository attribute and its validation in service are added to the project in the create method.

+++ src/main/java/com/ceciliasuarez/project/service/impl/CategoryServiceImpl.java
@@ -1,10 +1,10 @@
 package com.ceciliasuarez.project.service.impl;
 
 import com.ceciliasuarez.project.exceptions.DuplicateResourceException;
+import com.ceciliasuarez.project.exceptions.ResourceNotFoundException;
 import com.ceciliasuarez.project.model.Category;
 import com.ceciliasuarez.project.repository.ICategoryRepository;
 import com.ceciliasuarez.project.service.ICategoryService;
-import org.apache.coyote.BadRequestException;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,26 +20,35 @@ public class CategoryServiceImpl implements ICategoryService {
 
     @Override
     public List<Category> getAllCategory() {
+        logger.info("Categories list:");
         return categoryRepository.findAll();
+
     }
 
     @Override
     public Category createCategory(Category category) {
         logger.info("Saving new category...");
         List<Category> listCategories = getAllCategory();
-        for (Category existingcategory: listCategories) {
-            if (existingcategory.getName().equals(category.getName())){
-                logger.info("The new category could not be saved because a category with that name already exists.");
-                throw new DuplicateResourceException("A category with this name already exists.");
+        for (Category existingCategory: listCategories) {
+            if (existingCategory.getName().equals(category.getName())){
+                logger.info("Error when saving new category.");
+                throw new DuplicateResourceException("There is already a category with that name.");
             }
         }
-        logger.info("¡Category saved successfully!");
+        logger.info("Category saved successfully!");
         return categoryRepository.save(category);
 
     }
 
     @Override
     public void deleteCategory(Long id) {
+        logger.info("Removing category with id " + id);
+        if (!categoryRepository.findById(id).isPresent()) {
+            logger.info("Category removal failure.");
+            throw new ResourceNotFoundException("There is no category with the id " + id);
+        }
+        logger.info("Category removed successfully!");
         categoryRepository.deleteById(id);
     }
+
 }
